+++ src/routes/routing/loader/visualizeLoader.ts
@@ -1,8 +1,0 @@
-import type { LoaderFunctionArgs } from 'react-router-dom'
-
-import { prCore } from '@/createCore'
-
-export async function visualizeLoader({ request }: LoaderFunctionArgs) {
-  const { visualizeSurvey } = (await prCore).functions
-  return visualizeSurvey.loader({ requestUrl: request.url })
-}